malloc failed handled, to double check and final norminette
