+++ src/ast/parse_argument.c
@@ -17,7 +17,9 @@ t_ast	*parse_argument(t_ast *chd_ptr, t_parser *p)
 	t_ast	*word_ptr;
 	
 	word_ptr = NULL;
-	chd_ptr = create_treenode(chd_ptr);
+	chd_ptr = create_treenode(chd_ptr, p);
+	if (p->malloc_flag == 1)
+		return (NULL);
 	chd_ptr->type = AST_ARGUMENT;
 	if (!token_peek(p))
 	{
@@ -26,9 +28,13 @@ t_ast	*parse_argument(t_ast *chd_ptr, t_parser *p)
 	}
 	if (token_peek(p)->token == WORD)
 	{
-		word_ptr = create_treenode(word_ptr);
+		word_ptr = create_treenode(word_ptr, p);
+		if (p->malloc_flag == 1)
+			return (chd_ptr);
 		parse_word(word_ptr, p);
-		attach_treenode(chd_ptr, word_ptr);
+		attach_treenode(chd_ptr, word_ptr, p);
+		if (p->malloc_flag == 1)
+			free_treenode(word_ptr);
 	}
 	return (chd_ptr);
 }
