+++ src/ast/parse_pipe.c
@@ -6,7 +6,7 @@
 /*   By: jhor <jhor@student.42kl.edu.my>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/29 16:13:33 by jhor              #+#    #+#             */
-/*   Updated: 2025/10/22 23:36:46 by jhor             ###   ########.fr       */
+/*   Updated: 2025/10/23 15:01:38 by jhor             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -93,15 +93,3 @@ void	parse_pipeline(t_ast *root, t_parser *p)
 		parse_pipeline(root, p);
 }
 
-t_ast	*parsing(t_ast *node, t_token *token, t_parser *p)
-{
-	if (!token)
-		return (NULL);
-	node = init_ast(node, p, token);
-	parse_pipeline(node, p);
-	if (p->malloc_flag == 1)
-		ast_exit(token, node, p->result);
-	if (p->err_flag == 1)
-		return (NULL);
-	return (node);
-}
