+++ src/create_node.c
@@ -6,13 +6,13 @@
 /*   By: jhor <jhor@student.42kl.edu.my>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/09/14 17:50:35 by jhor              #+#    #+#             */
-/*   Updated: 2025/10/22 22:00:51 by jhor             ###   ########.fr       */
+/*   Updated: 2025/10/23 15:04:31 by jhor             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../minishell.h"
 
-char	*first_lexeme(t_token *token, int *i, char *start, size_t **end)
+char	*first_lexeme(t_token *token, size_t *i, char *start, size_t *end)
 {
 	while (*i < *end && start[*i])
 	{
@@ -27,10 +27,8 @@ t_token	*create_node(t_token *head, char *start, size_t end)
 {
 	t_token	*temp;
 	size_t	i;
-	int		flag;
 	
 	i = 0;
-	flag = 0;
 	temp = malloc(sizeof(t_token));
 	if (!temp)
 		return (NULL);
